feat: Create appropriate role verification

+++ backend/app/Http/Controllers/Api/V1/BookController.php
@@ -10,6 +10,7 @@ use App\Http\Resources\V1\BookResource;
 use App\Http\Resources\V1\BookCollection;
 use Illuminate\Http\Request;
 use App\Filters\V1\BookFilter;
+use Illuminate\Support\Facades\Gate;
 
 class BookController extends Controller
 {
@@ -18,6 +19,8 @@ class BookController extends Controller
      */
     public function index(Request $request)
     {
+        Gate::authorize('viewAny', Book::class);
+
         $filter = new BookFilter();
         $filterItems = $filter->transform($request); // [['column', 'operator', 'value']])
         
@@ -37,6 +40,9 @@ class BookController extends Controller
      */
     public function store(StoreBookRequest $request)
     {
+        Gate::authorize('create', Book::class);
+        
+
         $book = Book::create($request->except('category_ids'));
         $book->categories()->attach($request->category_ids);
         
@@ -48,6 +54,8 @@ class BookController extends Controller
      */
     public function show(Book $book)
     {
+        Gate::authorize('view', $book);
+
         $includeReservations = request()->query('includeReservations');
 
         if ($includeReservations) {
@@ -62,6 +70,8 @@ class BookController extends Controller
      */
     public function update(UpdateBookRequest $request, Book $book)
     {
+        Gate::authorize('update', $book);
+
         $book->update($request->except('category_ids'));
         if ($request->category_ids) {
         $book->categories()->sync($request->category_ids);
@@ -75,6 +85,8 @@ class BookController extends Controller
      */
     public function destroy(Book $book)
     {
+        Gate::authorize('delete', $book);
+        
         $book->categories()->detach();
         if ($book->reservations()->count() > 0 || $book->reservations()->count() > 0) {
             return response()->json(['message' => 'Book has reservations'], 409);
